requirements.txt
appdirs==1.4.4
distlib==0.3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
filelock==3.0.12
six==1.15.0
virtualenv==20.0.25
